+++ src/lantern/config.py
@@ -66,7 +66,7 @@ class Config:
         STORE_GITLAB_ENDPOINT: str
         STORE_GITLAB_TOKEN: str
         STORE_GITLAB_PROJECT_ID: str
-        STORE_GITLAB_BRANCH: str
+        STORE_GITLAB_DEFAULT_BRANCH: str
         STORE_GITLAB_CACHE_PATH: str
         TEMPLATES_CACHE_BUST_VALUE: str
         TEMPLATES_PLAUSIBLE_ID: str
@@ -100,7 +100,7 @@ class Config:
             "STORE_GITLAB_ENDPOINT": self.STORE_GITLAB_ENDPOINT,
             "STORE_GITLAB_TOKEN": self.STORE_GITLAB_TOKEN_SAFE,
             "STORE_GITLAB_PROJECT_ID": self.STORE_GITLAB_PROJECT_ID,
-            "STORE_GITLAB_BRANCH": self.STORE_GITLAB_BRANCH,
+            "STORE_GITLAB_DEFAULT_BRANCH": self.STORE_GITLAB_DEFAULT_BRANCH,
             "STORE_GITLAB_CACHE_PATH": str(self.STORE_GITLAB_CACHE_PATH.resolve()),
             "TEMPLATES_CACHE_BUST_VALUE": self.TEMPLATES_CACHE_BUST_VALUE,
             "TEMPLATES_PLAUSIBLE_ID": self.TEMPLATES_PLAUSIBLE_ID,
@@ -247,10 +247,10 @@ class Config:
             return self._env.str("PROJECT_ID")
 
     @property
-    def STORE_GITLAB_BRANCH(self) -> str:
+    def STORE_GITLAB_DEFAULT_BRANCH(self) -> str:
         """Branch name for GitLab store."""
         with self._env.prefixed(self._app_prefix), self._env.prefixed("STORE_GITLAB_"):
-            return self._env.str("BRANCH")
+            return self._env.str("DEFAULT_BRANCH")
 
     @property
     def STORE_GITLAB_CACHE_PATH(self) -> Path: