+++ tests/resources/stores/fake_records_store.py
@@ -2,7 +2,7 @@ import logging
 from copy import copy
 
 from lantern.models.record.revision import RecordRevision
-from lantern.stores.base import RecordNotFoundError, RecordsNotFoundError, StoreBase
+from lantern.stores.base import RecordNotFoundError, RecordsNotFoundError, StoreBase, StoreFrozenUnsupportedError
 from tests.resources.records.item_cat_checks import record as check
 from tests.resources.records.item_cat_collection_all import record as collection_all_supported
 from tests.resources.records.item_cat_collection_open_min import record as collection_open_min_required
@@ -40,10 +40,10 @@ class FakeRecordsStore(StoreBase):
     Termed 'fake' rather than 'test' to avoid confusion between testing a store, and a store used for that testing.
     """
 
-    def __init__(self, logger: logging.Logger, frozen: bool = False) -> None:
+    def __init__(self, logger: logging.Logger) -> None:
         self._logger = logger
         self._records: list[RecordRevision] = self._fake_records
-        self._frozen = frozen
+        self._frozen = False
 
     def __len__(self) -> int:
         """Count of records in store."""
@@ -119,3 +119,7 @@ class FakeRecordsStore(StoreBase):
             return self.select(file_identifiers={file_identifier})[0]
         except RecordsNotFoundError as e:
             raise RecordNotFoundError(file_identifier) from e
+
+    def freeze(self) -> None:
+        """Attempt to freeze store."""
+        raise StoreFrozenUnsupportedError() from None